refactor: restructure Arxiv MCP implementation
Reorganize Arxiv MCP code structure by moving from distributed capabilities to centralized implementation directory and unified server entry point.
