+++ Arxiv/.gitignore
@@ -28,3 +28,4 @@ dist/
 
 # Literature context files
 literature_context.md
+demo.py
