+++ Arxiv/src/arxiv_mcp.egg-info/SOURCES.txt
@@ -1,25 +1,23 @@
 README.md
 pyproject.toml
-src/arxiv_mcp/__init__.py
-src/arxiv_mcp/mcp_handlers.py
-src/arxiv_mcp/server.py
 src/arxiv_mcp.egg-info/PKG-INFO
 src/arxiv_mcp.egg-info/SOURCES.txt
 src/arxiv_mcp.egg-info/dependency_links.txt
 src/arxiv_mcp.egg-info/entry_points.txt
 src/arxiv_mcp.egg-info/requires.txt
 src/arxiv_mcp.egg-info/top_level.txt
-src/arxiv_mcp/capabilities/__init__.py
-src/arxiv_mcp/capabilities/arxiv_base.py
-src/arxiv_mcp/capabilities/category_search.py
-src/arxiv_mcp/capabilities/date_search.py
-src/arxiv_mcp/capabilities/export_utils.py
-src/arxiv_mcp/capabilities/paper_details.py
-src/arxiv_mcp/capabilities/pdf_download.py
-src/arxiv_mcp/capabilities/text_search.py
+src/implementation/__init__.py
+src/implementation/arxiv_base.py
+src/implementation/category_search.py
+src/implementation/date_search.py
+src/implementation/download_paper.py
+src/implementation/export_utils.py
+src/implementation/mcp_handlers.py
+src/implementation/paper_details.py
+src/implementation/text_search.py
 tests/test_category_search.py
+tests/test_download_paper.py
 tests/test_export_utils.py
 tests/test_integration.py
 tests/test_paper_details.py
-tests/test_pdf_download.py
 tests/test_text_search.py
