+++ Arxiv/src/arxiv_mcp.egg-info/entry_points.txt
@@ -1,2 +1,2 @@
 [console_scripts]
-arxiv-mcp = arxiv_mcp.server:main
+arxiv-mcp = server:main
