+++ Arxiv/src/arxiv_mcp.egg-info/top_level.txt
@@ -1,1 +1,1 @@
-arxiv_mcp
+implementation
