--- Arxiv/src/arxiv_mcp/capabilities/__init__.py
@@ -1,3 +1,0 @@
-"""
-ArXiv MCP capabilities.
-"""
